Restructuring the handlers and building an initial implementation for the transport connection handlers
