+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft.Data.SqlClient.csproj
@@ -920,6 +920,11 @@
     </Compile>
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\ConnectionHandlerContext.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\DataSourceParsingHandler.cs" />
+    <Compile Include="Microsoft\Data\SqlClientX\Handlers\NoSuitableHandlerFoundException.cs" />
+    <Compile Include="Microsoft\Data\SqlClientX\Handlers\ReturningHandler.cs" />
+    <Compile Include="Microsoft\Data\SqlClientX\Handlers\TransportCreation\NamedPipeTransportCreationHandler.cs" />
+    <Compile Include="Microsoft\Data\SqlClientX\Handlers\TransportCreation\SharedMemoryTransportCreationHandler.cs" />
+    <Compile Include="Microsoft\Data\SqlClientX\Handlers\TransportCreation\TcpTransportCreationHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\IO\ITdsReadStream.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\IO\ITdsStream.cs" />
@@ -927,9 +932,7 @@
     <Compile Include="Microsoft\Data\SqlClientX\IO\TdsReadStream.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\IO\TdsStreamPacketType.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\IO\TdsStream.cs" />
-    <Compile Include="Microsoft\Data\SqlClientX\Handlers\HandlerRequest.cs" />
-    <Compile Include="Microsoft\Data\SqlClientX\Handlers\HandlerRequestType.cs" />
-    <Compile Include="Microsoft\Data\SqlClientX\Handlers\IHandler.cs" />
+    <Compile Include="Microsoft\Data\SqlClientX\Handlers\ContextHandler.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\TransportCreation\IpAddressVersionComparer.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\TransportCreation\TransportCreationHandler.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\IO\TdsWriteStream.cs" />
